엔티티수정
